services/backend/requirements.txt
coverage==4.5.2
flake8==3.6.0
Flask==1.0.2
flask-bcrypt==0.7.1
flask-cors==3.0.7
flask-debugtoolbar==0.10.1
flask-migrate==2.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
pyjwt==1.7.1
pytest==3.10.0